Transfer data input checks refactoring & received value calculation fix
Refactors the validation of transfer data input
Fixes bug related to the calculation of received value for transfers

+++ BudgetManager/utils/data_insertion/TransferCheckStrategy.cs
@@ -13,7 +13,7 @@ namespace BudgetManager.non_mvc {
     class TransferCheckStrategy : DataInsertionCheckStrategy {
         private String accountBalanceCheckProcedure = "can_perform_requested_transfer";
     
-        public DataCheckResponse performCheck(QueryData inputData, string selectedItemName, int valueToInsert) {
+        public DataCheckResponse performCheck(QueryData inputData, string selectedItemName, double valueToInsert) {
             DataCheckResponse dataCheckResponse = new DataCheckResponse();
 
             if (checkAvailableBalance(inputData, valueToInsert) == 0) {
@@ -31,7 +31,7 @@ namespace BudgetManager.non_mvc {
 
         //Method that checks if the transferred amount is lower/equal to the balance of the specified account, using a database stored procedure
         //The stored procedure returns 1 (true) if the transfer can be performed and 0 (false) otherwise
-        private int checkAvailableBalance(QueryData inputData, int transferValue) {
+        private int checkAvailableBalance(QueryData inputData, double transferValue) {
             MySqlParameter checkResultOutput = null;
             MySqlParameter accountBalanceOutput = null;
 
@@ -51,7 +51,7 @@ namespace BudgetManager.non_mvc {
                 checkResultOutput = new MySqlParameter("p_result", MySqlDbType.Int32);
                 checkResultOutput.Direction = System.Data.ParameterDirection.Output;
 
-                accountBalanceOutput = new MySqlParameter("p_account_balance", MySqlDbType.Int32);
+                accountBalanceOutput = new MySqlParameter("p_account_balance", MySqlDbType.Double);
                 accountBalanceOutput.Direction = System.Data.ParameterDirection.Output;
 
                 //Adds parameters to the command
@@ -82,7 +82,7 @@ namespace BudgetManager.non_mvc {
 
         }
 
-        public DataCheckResponse performCheck(QueryData paramContainer, String selectedItemName, double valueToInsert) {
+        public DataCheckResponse performCheck(QueryData paramContainer, String selectedItemName, int valueToInsert) {
             throw new NotImplementedException();
         }
